#27 [MOD] 이웃목록 페이지 프래그먼트로 변경
