+++ app/src/main/res/layout/fragment_neighbors.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.profile.NeighborsActivity">
+    tools:context=".presentation.profile.NeighborsFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_modify_header"
